third_party/WebKit/LayoutTests/webmidi/requestmidiaccess-fail-expected.txt

Tests navigator.requestMIDIAccess failure.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS errorName is 'InvalidStateError'
PASS requestMIDIAccess() fail as expected.
PASS successfullyParsed is true

TEST COMPLETE
